deployment/cloud-run/requirements.txt
google-cloud-bigquery[bqstorage]==3.27.0
google-cloud-secret-manager==2.21.1
duckdb==1.4.1
pyarrow==19.0.0
websockets==14.1
